refactor(thread_ops): split low-cohesion module into focused submodules
Extract 18 functions from src/agent/thread_ops.rs into 4 new/expanded
submodules organized by responsibility:

- control.rs (8 functions): Thread lifecycle state transitions
  - process_undo, process_redo, process_interrupt, process_compact
  - process_clear, process_new_thread, process_switch_thread, process_resume

- persistence.rs (+3 methods): DB persistence for messages and tool calls
  - persist_user_message, persist_assistant_response, persist_tool_calls

- turn_execution.rs (1 function): User turn lifecycle orchestration
  - process_user_input (~360 lines)

- hydration.rs (2 functions): Thread hydration from database
  - hydrate_and_resolve_session_thread, maybe_hydrate_thread

The root file now contains only message-handling orchestration (4 items):
- store_extracted_documents (free function)
- handle_message (main entry point)
- check_auth_mode_intercept (private helper)
- set_tool_context_for_message (private helper)

This reduces the responsibility count below CodeScene's threshold of 3,
addressing the Low Cohesion biomarker while preserving all existing
behaviour and public APIs.

Closes #122
