+++ src/agent/thread_ops/control.rs
@@ -1,0 +1,235 @@
+//! Thread control command handlers.
+//!
+//! Contains handlers for thread lifecycle state transitions:
+//! - Undo/redo operations
+//! - Interrupt processing
+//! - Context compaction
+//! - Thread clearing
+//! - New thread creation
+//! - Thread switching
+//! - Resume from checkpoint
+
+use std::sync::Arc;
+
+use tokio::sync::Mutex;
+use uuid::Uuid;
+
+use crate::agent::Agent;
+use crate::agent::compaction::ContextCompactor;
+use crate::agent::session::{Session, ThreadState};
+use crate::agent::submission::SubmissionResult;
+use crate::error::Error;
+
+impl Agent {
+    pub(super) async fn process_undo(
+        &self,
+        session: Arc<Mutex<Session>>,
+        thread_id: Uuid,
+    ) -> Result<SubmissionResult, Error> {
+        let undo_mgr = self.session_manager.get_undo_manager(thread_id).await;
+        let mut mgr = undo_mgr.lock().await;
+
+        if !mgr.can_undo() {
+            return Ok(SubmissionResult::ok_with_message("Nothing to undo."));
+        }
+
+        let mut sess = session.lock().await;
+        let thread = sess
+            .threads
+            .get_mut(&thread_id)
+            .ok_or_else(|| Error::from(crate::error::JobError::NotFound { id: thread_id }))?;
+
+        // Save current state to redo, get previous checkpoint
+        let current_messages = thread.messages();
+        let current_turn = thread.turn_number();
+
+        if let Some(checkpoint) = mgr.undo(current_turn, current_messages) {
+            // Extract values before consuming the reference
+            let turn_number = checkpoint.turn_number;
+            let messages = checkpoint.messages.clone();
+            let undo_count = mgr.undo_count();
+            // Restore thread from checkpoint
+            thread.restore_from_messages(messages);
+            Ok(SubmissionResult::ok_with_message(format!(
+                "Undone to turn {}. {} undo(s) remaining.",
+                turn_number, undo_count
+            )))
+        } else {
+            Ok(SubmissionResult::error("Undo failed."))
+        }
+    }
+
+    pub(super) async fn process_redo(
+        &self,
+        session: Arc<Mutex<Session>>,
+        thread_id: Uuid,
+    ) -> Result<SubmissionResult, Error> {
+        let undo_mgr = self.session_manager.get_undo_manager(thread_id).await;
+        let mut mgr = undo_mgr.lock().await;
+
+        if !mgr.can_redo() {
+            return Ok(SubmissionResult::ok_with_message("Nothing to redo."));
+        }
+
+        let mut sess = session.lock().await;
+        let thread = sess
+            .threads
+            .get_mut(&thread_id)
+            .ok_or_else(|| Error::from(crate::error::JobError::NotFound { id: thread_id }))?;
+
+        let current_messages = thread.messages();
+        let current_turn = thread.turn_number();
+
+        if let Some(checkpoint) = mgr.redo(current_turn, current_messages) {
+            thread.restore_from_messages(checkpoint.messages);
+            Ok(SubmissionResult::ok_with_message(format!(
+                "Redone to turn {}.",
+                checkpoint.turn_number
+            )))
+        } else {
+            Ok(SubmissionResult::error("Redo failed."))
+        }
+    }
+
+    pub(super) async fn process_interrupt(
+        &self,
+        session: Arc<Mutex<Session>>,
+        thread_id: Uuid,
+    ) -> Result<SubmissionResult, Error> {
+        let mut sess = session.lock().await;
+        let thread = sess
+            .threads
+            .get_mut(&thread_id)
+            .ok_or_else(|| Error::from(crate::error::JobError::NotFound { id: thread_id }))?;
+
+        match thread.state {
+            ThreadState::Processing | ThreadState::AwaitingApproval => {
+                thread.interrupt();
+                Ok(SubmissionResult::ok_with_message("Interrupted."))
+            }
+            _ => Ok(SubmissionResult::ok_with_message("Nothing to interrupt.")),
+        }
+    }
+
+    pub(super) async fn process_compact(
+        &self,
+        session: Arc<Mutex<Session>>,
+        thread_id: Uuid,
+    ) -> Result<SubmissionResult, Error> {
+        let mut sess = session.lock().await;
+        let thread = sess
+            .threads
+            .get_mut(&thread_id)
+            .ok_or_else(|| Error::from(crate::error::JobError::NotFound { id: thread_id }))?;
+
+        let messages = thread.messages();
+        let usage = self.context_monitor.usage_percent(&messages);
+        let strategy = self
+            .context_monitor
+            .suggest_compaction(&messages)
+            .unwrap_or(
+                crate::agent::context_monitor::CompactionStrategy::Summarize { keep_recent: 5 },
+            );
+
+        let compactor = ContextCompactor::new(self.llm().clone());
+        match compactor
+            .compact(thread, strategy, self.workspace().map(|w| w.as_ref()))
+            .await
+        {
+            Ok(result) => {
+                let mut msg = format!(
+                    "Compacted: {} turns removed, {} → {} tokens (was {:.1}% full)",
+                    result.turns_removed, result.tokens_before, result.tokens_after, usage
+                );
+                if result.summary_written {
+                    msg.push_str(", summary saved to workspace");
+                }
+                Ok(SubmissionResult::ok_with_message(msg))
+            }
+            Err(e) => Ok(SubmissionResult::error(format!("Compaction failed: {}", e))),
+        }
+    }
+
+    pub(super) async fn process_clear(
+        &self,
+        session: Arc<Mutex<Session>>,
+        thread_id: Uuid,
+    ) -> Result<SubmissionResult, Error> {
+        let mut sess = session.lock().await;
+        let thread = sess
+            .threads
+            .get_mut(&thread_id)
+            .ok_or_else(|| Error::from(crate::error::JobError::NotFound { id: thread_id }))?;
+        thread.turns.clear();
+        thread.state = ThreadState::Idle;
+
+        // Clear undo history too
+        let undo_mgr = self.session_manager.get_undo_manager(thread_id).await;
+        undo_mgr.lock().await.clear();
+
+        Ok(SubmissionResult::ok_with_message("Thread cleared."))
+    }
+
+    pub(super) async fn process_new_thread(
+        &self,
+        message: &crate::channels::IncomingMessage,
+    ) -> Result<SubmissionResult, Error> {
+        let session = self
+            .session_manager
+            .get_or_create_session(&message.user_id)
+            .await;
+        let mut sess = session.lock().await;
+        let thread = sess.create_thread();
+        let thread_id = thread.id;
+        Ok(SubmissionResult::ok_with_message(format!(
+            "New thread: {}",
+            thread_id
+        )))
+    }
+
+    pub(super) async fn process_switch_thread(
+        &self,
+        message: &crate::channels::IncomingMessage,
+        target_thread_id: Uuid,
+    ) -> Result<SubmissionResult, Error> {
+        let session = self
+            .session_manager
+            .get_or_create_session(&message.user_id)
+            .await;
+        let mut sess = session.lock().await;
+
+        if sess.switch_thread(target_thread_id) {
+            Ok(SubmissionResult::ok_with_message(format!(
+                "Switched to thread {}",
+                target_thread_id
+            )))
+        } else {
+            Ok(SubmissionResult::error("Thread not found."))
+        }
+    }
+
+    pub(super) async fn process_resume(
+        &self,
+        session: Arc<Mutex<Session>>,
+        thread_id: Uuid,
+        checkpoint_id: Uuid,
+    ) -> Result<SubmissionResult, Error> {
+        let undo_mgr = self.session_manager.get_undo_manager(thread_id).await;
+        let mut mgr = undo_mgr.lock().await;
+
+        if let Some(checkpoint) = mgr.restore(checkpoint_id) {
+            let mut sess = session.lock().await;
+            let thread = sess
+                .threads
+                .get_mut(&thread_id)
+                .ok_or_else(|| Error::from(crate::error::JobError::NotFound { id: thread_id }))?;
+            thread.restore_from_messages(checkpoint.messages);
+            Ok(SubmissionResult::ok_with_message(format!(
+                "Resumed from checkpoint: {}",
+                checkpoint.description
+            )))
+        } else {
+            Ok(SubmissionResult::error("Checkpoint not found."))
+        }
+    }
+}
