+++ src/agent/thread_ops/turn_execution.rs
@@ -1,0 +1,388 @@
+//! User turn execution and agentic loop orchestration.
+//!
+//! Handles the full lifecycle of a user input turn:
+//! - Thread state validation
+//! - Safety checks (input validation, policy, secrets)
+//! - Command routing
+//! - Auto-compaction
+//! - Undo checkpointing
+//! - Attachment augmentation
+//! - Agentic loop execution
+//! - Response persistence
+
+use std::sync::Arc;
+
+use tokio::sync::Mutex;
+use uuid::Uuid;
+
+use crate::agent::Agent;
+use crate::agent::compaction::ContextCompactor;
+use crate::agent::dispatcher::AgenticLoopResult;
+use crate::agent::session::{Session, ThreadState};
+use crate::agent::submission::SubmissionResult;
+use crate::channels::{IncomingMessage, StatusUpdate};
+use crate::error::Error;
+
+impl Agent {
+    pub(super) async fn process_user_input(
+        &self,
+        message: &IncomingMessage,
+        session: Arc<Mutex<Session>>,
+        thread_id: Uuid,
+        content: &str,
+    ) -> Result<SubmissionResult, Error> {
+        tracing::debug!(
+            message_id = %message.id,
+            thread_id = %thread_id,
+            content_len = content.len(),
+            "Processing user input"
+        );
+
+        // First check thread state without holding lock during I/O
+        let thread_state = {
+            let sess = session.lock().await;
+            let thread = sess
+                .threads
+                .get(&thread_id)
+                .ok_or_else(|| Error::from(crate::error::JobError::NotFound { id: thread_id }))?;
+            thread.state
+        };
+
+        tracing::debug!(
+            message_id = %message.id,
+            thread_id = %thread_id,
+            thread_state = ?thread_state,
+            "Checked thread state"
+        );
+
+        // Check thread state
+        match thread_state {
+            ThreadState::Processing => {
+                tracing::warn!(
+                    message_id = %message.id,
+                    thread_id = %thread_id,
+                    "Thread is processing, rejecting new input"
+                );
+                return Ok(SubmissionResult::error(
+                    "Turn in progress. Use /interrupt to cancel.",
+                ));
+            }
+            ThreadState::AwaitingApproval => {
+                tracing::warn!(
+                    message_id = %message.id,
+                    thread_id = %thread_id,
+                    "Thread awaiting approval, rejecting new input"
+                );
+                return Ok(SubmissionResult::error(
+                    "Waiting for approval. Use /interrupt to cancel.",
+                ));
+            }
+            ThreadState::Completed => {
+                tracing::warn!(
+                    message_id = %message.id,
+                    thread_id = %thread_id,
+                    "Thread completed, rejecting new input"
+                );
+                return Ok(SubmissionResult::error(
+                    "Thread completed. Use /thread new.",
+                ));
+            }
+            ThreadState::Idle | ThreadState::Interrupted => {
+                // Can proceed
+            }
+        }
+
+        // Safety validation for user input
+        let validation = self.safety().validate_input(content);
+        if !validation.is_valid {
+            let details = validation
+                .errors
+                .iter()
+                .map(|e| format!("{}: {}", e.field, e.message))
+                .collect::<Vec<_>>()
+                .join("; ");
+            return Ok(SubmissionResult::error(format!(
+                "Input rejected by safety validation: {}",
+                details
+            )));
+        }
+
+        let violations = self.safety().check_policy(content);
+        if violations
+            .iter()
+            .any(|rule| rule.action == crate::safety::PolicyAction::Block)
+        {
+            return Ok(SubmissionResult::error("Input rejected by safety policy."));
+        }
+
+        // Scan inbound messages for secrets (API keys, tokens).
+        // Catching them here prevents the LLM from echoing them back, which
+        // would trigger the outbound leak detector and create error loops.
+        if let Some(warning) = self.safety().scan_inbound_for_secrets(content) {
+            tracing::warn!(
+                user = %message.user_id,
+                channel = %message.channel,
+                "Inbound message blocked: contains leaked secret"
+            );
+            return Ok(SubmissionResult::error(warning));
+        }
+
+        // Handle explicit commands (starting with /) directly
+        // Everything else goes through the normal agentic loop with tools
+        let temp_message = IncomingMessage {
+            content: content.to_string(),
+            ..message.clone()
+        };
+
+        if let Some(intent) = self.router.route_command(&temp_message) {
+            // Explicit command like /status, /job, /list - handle directly
+            return self.handle_job_or_command(intent, message).await;
+        }
+
+        // Natural language goes through the agentic loop
+        // Job tools (create_job, list_jobs, etc.) are in the tool registry
+
+        // Auto-compact if needed BEFORE adding new turn
+        {
+            let mut sess = session.lock().await;
+            let thread = sess
+                .threads
+                .get_mut(&thread_id)
+                .ok_or_else(|| Error::from(crate::error::JobError::NotFound { id: thread_id }))?;
+
+            let messages = thread.messages();
+            if let Some(strategy) = self.context_monitor.suggest_compaction(&messages) {
+                let pct = self.context_monitor.usage_percent(&messages);
+                tracing::info!("Context at {:.1}% capacity, auto-compacting", pct);
+
+                // Notify the user that compaction is happening
+                let _ = self
+                    .channels
+                    .send_status(
+                        &message.channel,
+                        StatusUpdate::Status(format!(
+                            "Context at {:.0}% capacity, compacting...",
+                            pct
+                        )),
+                        &message.metadata,
+                    )
+                    .await;
+
+                let compactor = ContextCompactor::new(self.llm().clone());
+                if let Err(e) = compactor
+                    .compact(thread, strategy, self.workspace().map(|w| w.as_ref()))
+                    .await
+                {
+                    tracing::warn!("Auto-compaction failed: {}", e);
+                }
+            }
+        }
+
+        // Create checkpoint before turn
+        let undo_mgr = self.session_manager.get_undo_manager(thread_id).await;
+        {
+            let sess = session.lock().await;
+            let thread = sess
+                .threads
+                .get(&thread_id)
+                .ok_or_else(|| Error::from(crate::error::JobError::NotFound { id: thread_id }))?;
+
+            let mut mgr = undo_mgr.lock().await;
+            mgr.checkpoint(
+                thread.turn_number(),
+                thread.messages(),
+                format!("Before turn {}", thread.turn_number()),
+            );
+        }
+
+        // Augment content with attachment context (transcripts, metadata, images)
+        let augmented =
+            crate::agent::attachments::augment_with_attachments(content, &message.attachments);
+        let (effective_content, image_parts) = match &augmented {
+            Some(result) => (result.text.as_str(), result.image_parts.clone()),
+            None => (content, Vec::new()),
+        };
+
+        // Start the turn and get messages
+        let turn_messages = {
+            let mut sess = session.lock().await;
+            let thread = sess
+                .threads
+                .get_mut(&thread_id)
+                .ok_or_else(|| Error::from(crate::error::JobError::NotFound { id: thread_id }))?;
+            let turn = thread.start_turn(effective_content);
+            turn.image_content_parts = image_parts;
+            thread.messages()
+        };
+
+        // Persist user message to DB immediately so it survives crashes
+        tracing::debug!(
+            message_id = %message.id,
+            thread_id = %thread_id,
+            "Persisting user message to DB"
+        );
+        self.persist_user_message(thread_id, &message.user_id, effective_content)
+            .await;
+
+        tracing::debug!(
+            message_id = %message.id,
+            thread_id = %thread_id,
+            "User message persisted, starting agentic loop"
+        );
+
+        // Send thinking status
+        let _ = self
+            .channels
+            .send_status(
+                &message.channel,
+                StatusUpdate::Thinking("Processing...".into()),
+                &message.metadata,
+            )
+            .await;
+
+        // Run the agentic tool execution loop
+        let result = self
+            .run_agentic_loop(message, session.clone(), thread_id, turn_messages)
+            .await;
+
+        // Re-acquire lock and check if interrupted
+        let interrupted = {
+            let mut sess = session.lock().await;
+            let thread = sess
+                .threads
+                .get_mut(&thread_id)
+                .ok_or_else(|| Error::from(crate::error::JobError::NotFound { id: thread_id }))?;
+            thread.state == ThreadState::Interrupted
+        };
+        if interrupted {
+            let _ = self
+                .channels
+                .send_status(
+                    &message.channel,
+                    StatusUpdate::Status("Interrupted".into()),
+                    &message.metadata,
+                )
+                .await;
+            return Ok(SubmissionResult::Interrupted);
+        }
+
+        // Re-acquire lock for processing result
+        let mut sess = session.lock().await;
+        let thread = sess
+            .threads
+            .get_mut(&thread_id)
+            .ok_or_else(|| Error::from(crate::error::JobError::NotFound { id: thread_id }))?;
+
+        // Complete, fail, or request approval
+        match result {
+            Ok(AgenticLoopResult::Response(response)) => {
+                // Drop the session lock before running the response transform hook
+                drop(sess);
+
+                // Hook: TransformResponse — allow hooks to modify or reject the final response
+                let response = {
+                    let event = crate::hooks::HookEvent::ResponseTransform {
+                        user_id: message.user_id.clone(),
+                        thread_id: thread_id.to_string(),
+                        response: response.clone(),
+                    };
+                    match self.hooks().run(&event).await {
+                        Err(crate::hooks::HookError::Rejected { reason }) => {
+                            format!("[Response filtered: {}]", reason)
+                        }
+                        Ok(crate::hooks::HookOutcome::Reject { reason }) => {
+                            format!("[Response filtered: {}]", reason)
+                        }
+                        Err(err) => {
+                            tracing::warn!("TransformResponse hook failed open: {}", err);
+                            response
+                        }
+                        Ok(crate::hooks::HookOutcome::Continue {
+                            modified: Some(new_response),
+                        }) => new_response,
+                        _ => response, // fail-open: use original
+                    }
+                };
+
+                // Re-acquire lock to complete turn and snapshot data
+                let completion = {
+                    let mut sess = session.lock().await;
+                    let thread = sess.threads.get_mut(&thread_id).ok_or_else(|| {
+                        Error::from(crate::error::JobError::NotFound { id: thread_id })
+                    })?;
+                    if thread.state == ThreadState::Interrupted {
+                        None
+                    } else {
+                        thread.complete_turn(&response);
+                        Some(
+                            thread
+                                .turns
+                                .last()
+                                .map(|t| (t.turn_number, t.tool_calls.clone()))
+                                .unwrap_or_default(),
+                        )
+                    }
+                };
+                let Some((turn_number, tool_calls)) = completion else {
+                    let _ = self
+                        .channels
+                        .send_status(
+                            &message.channel,
+                            StatusUpdate::Status("Interrupted".into()),
+                            &message.metadata,
+                        )
+                        .await;
+                    return Ok(SubmissionResult::Interrupted);
+                };
+                // Lock is dropped here at end of block
+
+                let _ = self
+                    .channels
+                    .send_status(
+                        &message.channel,
+                        StatusUpdate::Status("Done".into()),
+                        &message.metadata,
+                    )
+                    .await;
+
+                // Persist tool calls then assistant response (user message already persisted at turn start)
+                self.persist_tool_calls(thread_id, &message.user_id, turn_number, &tool_calls)
+                    .await;
+                self.persist_assistant_response(thread_id, &message.user_id, &response)
+                    .await;
+
+                Ok(SubmissionResult::response(response))
+            }
+            Ok(AgenticLoopResult::NeedApproval { pending }) => {
+                // Store pending approval in thread and update state
+                let request_id = pending.request_id;
+                let tool_name = pending.tool_name.clone();
+                let description = pending.description.clone();
+                let parameters = pending.display_parameters.clone();
+                thread.await_approval(pending);
+                // Drop the session lock before async operations
+                drop(sess);
+
+                let _ = self
+                    .channels
+                    .send_status(
+                        &message.channel,
+                        StatusUpdate::Status("Awaiting approval".into()),
+                        &message.metadata,
+                    )
+                    .await;
+                Ok(SubmissionResult::NeedApproval {
+                    request_id,
+                    tool_name,
+                    description,
+                    parameters,
+                })
+            }
+            Err(e) => {
+                thread.fail_turn(e.to_string());
+                // User message already persisted at turn start; nothing else to save
+                Ok(SubmissionResult::error(e.to_string()))
+            }
+        }
+    }
+}
